+++ packages/runtime/src/axios.ts
@@ -1,0 +1,7 @@
+import axios from 'axios';
+
+export type { AxiosRequestConfig, CancelTokenStatic, CancelStatic, AxiosResponse, AxiosError } from 'axios';
+
+export {
+  axios
+};