+++ packages/runtime/src/iceStore.ts
@@ -1,3 +1,6 @@
 /* for plugin-store  */
+import { createStore } from '@ice/store';
 
 export type { Models, IcestoreRootState, IcestoreDispatch } from '@ice/store';
+
+export { createStore };