+++ packages/runtime/src/pathToRegexp.ts
@@ -1,0 +1,5 @@
+import * as pathToRegexp from 'path-to-regexp';
+
+export {
+  pathToRegexp
+};