+++ packages/runtime/src/reactRouter.ts
@@ -1,0 +1,5 @@
+import { withRouter } from 'react-router';
+
+export {
+  withRouter
+};